+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -314,7 +314,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KS7AZ80TWMWJ1JQYV4NFBS6W"
+        "01KSETWJ935JNEZST0T697ZJNN"
     )
     r"""The unique identifier of the data source"""