+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KS7AZ80SADYVZ3M22FYHRRD9"
+        "01KSETWJ91GPQ3GJSW9T5020JW"
     )
     r"""The unique identifier of the data source"""